+++ core/src/hooks.cpp
@@ -4,7 +4,10 @@
 
 #include "core/log.hpp"
 #include "core/state.hpp"
+#include "core/system.hpp"
 #include "core/xinput.hpp"
+
+#include <filesystem>
 
 #include <windows.h>
 #include <MinHook.h>
@@ -28,24 +31,47 @@ GetProcAddress_t g_origGetProcAddress{};
 
 NTSTATUS NTAPI detourLdrLoadDll(PWSTR DllPath, PULONG DllCharacteristics, PUNICODE_STRING DllName, PVOID* DllHandle) {
     const std::wstring_view libName{DllName->Buffer, DllName->Length / sizeof(DllName->Buffer[0])};
-    const std::filesystem::path libPath{libName};
+    auto libPath{std::filesystem::path{libName}.lexically_normal()};
     log("LdrLoadDll: ", charStringFromChar8String(libPath.u8string()));
     if (libPath.filename() == "gamedll_ph_x64_rwe.dll") {
         log("Game DLL is being loaded.");
         auto result{g_origLdrLoadDll(DllPath, DllCharacteristics, DllName, DllHandle)};
         AppState::get().notifyOnGameDllLoaded();
         return result;
-    } else if (libPath.filename() == "D3D12Core.dll") {
-        log("Redirecting loading of D3D12Core.");
-        const auto newPath{(getMainBinDir() / libPath.filename()).wstring()};
-        const auto newLengthInBytes{static_cast<USHORT>(newPath.size() * sizeof(DllName->Buffer[0]))};
-        UNICODE_STRING newPathUnicode{
-            .Length = newLengthInBytes,
-            .MaximumLength = newLengthInBytes,
-            .Buffer = const_cast<wchar_t*>(newPath.c_str())};
-        return g_origLdrLoadDll(DllPath, DllCharacteristics, &newPathUnicode, DllHandle);
     }
-    return g_origLdrLoadDll(DllPath, DllCharacteristics, DllName, DllHandle);
+    // Try to load DLL with provided path
+    auto result{g_origLdrLoadDll(DllPath, DllCharacteristics, DllName, DllHandle)};
+    if (result == 0) {
+        return result;
+    }
+    log("DLL failed to load: ", charStringFromChar8String(libPath.u8string()));
+    if (libPath.has_root_path()) {
+        // Might be rooted in the wrong directory
+        try {
+            const auto canonicalLibPath{std::filesystem::weakly_canonical(libPath)};
+            const auto relPath{std::filesystem::relative(libPath, getExeDir())};
+            libPath = getMainBinDir() / relPath;
+        } catch (const std::filesystem::filesystem_error& ex) {
+            log("Filesystem error: ", ex.what());
+            return result;
+        }
+    } else {
+        // Non-rooted might be in main game dir
+        libPath = getMainBinDir() / libPath;
+    }
+    log("Attempting to redirect to DLL: ", charStringFromChar8String(libPath.u8string()));
+    const auto newPath{(getMainBinDir() / libPath.filename()).wstring()};
+    const auto newLengthInBytes{static_cast<USHORT>(newPath.size() * sizeof(DllName->Buffer[0]))};
+    UNICODE_STRING newPathUnicode{
+        .Length = newLengthInBytes,
+        .MaximumLength = newLengthInBytes,
+        .Buffer = const_cast<wchar_t*>(newPath.c_str())};
+    result = g_origLdrLoadDll(DllPath, DllCharacteristics, &newPathUnicode, DllHandle);
+    if (result == 0) {
+        return result;
+    }
+    log("DLL still failed to load: ", charStringFromChar8String(libPath.u8string()));
+    return result;
 }
 
 FARPROC WINAPI detourGetProcAddress(HMODULE hModule, LPCSTR lpProcName) {